+++ section/route/route.js
@@ -2,7 +2,7 @@
 function route() {
     console.log("user function executed");
 
-    fetch('http://65.0.32.172/routeData')
+    fetch('http://localhost:3000/routeData')
         .then(response => {
             if (!response.ok) {
                 throw new Error('Network response was not ok');
@@ -23,6 +23,7 @@ function populateTable(data) {
     tbody.innerHTML = ''; // Clear existing rows
     var columnsToDisplay = ['route_name', 'details','mobile_no'];
     var counter = 1;
+    var isAdmin = JSON.parse(localStorage.getItem('sessionData'))[0].usertype === 'Admin';
     data.forEach(function(item) {
         var row = tbody.insertRow();
         var cell = row.insertCell();
@@ -32,32 +33,32 @@ function populateTable(data) {
             cell.textContent = item[key];
         });
 
-         // Add Edit button
-         var editCell = row.insertCell();
-         var editButton = document.createElement('button');
-         editButton.className = 'button edit-button';
-         var editLink = document.createElement('a');
-         editLink.textContent = 'Edit';
-         editButton.appendChild(editLink);
-         editLink.textContent = 'Edit';
-        editLink.style.color = 'white'; // Set text color to white
-        editLink.style.textDecoration = 'none'; // Remove underline
-
-
-        editButton.addEventListener('click', function() {
-            editRoute(item); // Pass the user data to the edit function
-        });
-        editCell.appendChild(editButton);
+        // Add Edit button if user is admin
+        if (isAdmin) {
+            var editCell = row.insertCell();
+            var editButton = document.createElement('button');
+            editButton.className = 'button edit-button';
+            var editLink = document.createElement('a');
+            editLink.href = '../route/update_route.html'; // Edit link destination
+            editLink.textContent = 'Edit';
+            editButton.appendChild(editLink);
+            editButton.addEventListener('click', function() {
+                editRoute(item); // Pass the user data to the edit function
+            });
+            editCell.appendChild(editButton);
+        }
  
-         // Add Delete button
-         var deleteCell = row.insertCell();
-         var deleteButton = document.createElement('button');
-         deleteButton.className = 'button delete-button';
-         deleteButton.textContent = 'Delete';
-         deleteButton.addEventListener('click', function() {
-            deleteRoute(item.id); // Pass the user id to the delete function
-        });
-        deleteCell.appendChild(deleteButton);
+        // Add Delete button if user is admin
+        if (isAdmin) {
+            var deleteCell = row.insertCell();
+            var deleteButton = document.createElement('button');
+            deleteButton.className = 'button delete-button';
+            deleteButton.textContent = 'Delete';
+            deleteButton.addEventListener('click', function() {
+                deleteRoute(item.id); // Pass the user id to the delete function
+            });
+            deleteCell.appendChild(deleteButton);
+        }
     });
 }
 
@@ -81,7 +82,7 @@ function editRoute(user) {
 
 function deleteRoute(userId) {
     // Perform delete operation based on userId
-    fetch('http://65.0.32.172/routeData/deleterouteId/' + userId, {
+    fetch('http://localhost:3000/routeData/deleterouteId/' + userId, {
         method: 'DELETE'
     })
     .then(response => {
